bugfix substring removal
In the original rstrip was used to remove '/jsonrpc' from url. Thus rstrip removes all characters on the end of jsonrpc_url in the set [/, j, s, o, n, r, p, c].
When the baseurl contains one of these characters at the end, these will also be removed, resulting in a wrong url.